+++ wikimap/Embeddings/vector_ops.hpp
@@ -12,22 +12,22 @@ namespace vec {
 
 template<class Target, class Source>
 inline void add_to(Target&& target, const Source& source) {
-    for (int i = 0; i < source.size(); ++i) {
+    for (size_t i = 0; i < source.size(); ++i) {
         target[i] += source[i];
     }
 }
 
 template<class Target>
 inline void fill_with_zeros(Target&& target) {
-    for (int i = 0; i < target.size(); ++i) {
+    for (size_t i = 0; i < target.size(); ++i) {
         target[i] = 0;
     }
 }
 
 template<class Source1, class Source2>
-inline double dot_product(const Source1& s1, const Source2& s2) {
-    double res = 0.;
-    for (int i = 0; i < s1.size(); ++i) {
+inline float dot_product(const Source1& s1, const Source2& s2) {
+    float res = 0.;
+    for (size_t i = 0; i < s1.size(); ++i) {
         res += s1[i] * s2[i];
     }
     return res;
@@ -36,11 +36,11 @@ inline double dot_product(const Source1& s1, const Source2& s2) {
 template<class Target>
 inline void normalize(Target&& target) {
     double sum = 0.;
-    for (int i = 0; i < target.size(); ++i) {
+    for (size_t i = 0; i < target.size(); ++i) {
         sum += target[i] * target[i];
     }
     double root = sqrt(sum);
-    for (int i = 0; i < target.size(); ++i) {
+    for (size_t i = 0; i < target.size(); ++i) {
         target[i] /= root;
     }
 }
